Expose gauss_legendre_tbl publicly as gauss_legendre_roots
